地面との判定をレイ判定に
もうちょっと当たり判定を調整。

+++ Src/Framework/Shader/StandardShader/KdStandardShader.h
@@ -51,7 +51,7 @@ public:
 		float			_blank[3] = { 0.0f, 0.0f ,0.0f };
 	};
 
-	// 定数バッファ(ボーン単位更新 : スキンメッシュ対応)
+	// 定数バッファ(ボーン単位更新：スキンメッシュ対応)
 	struct cbBone
 	{
 		Math::Matrix mBones[300];
@@ -209,7 +209,7 @@ private:
 	// スキンメッシュ有効かどうか(スキンメッシュ対応)
 	void SetIsSkinMeshObj(bool isSkinMeshObj)
 	{
-		if (m_cb0_Obj.Work().IsSkinMeshObj != isSkinMeshObj)
+		if (m_cb0_Obj.Work().IsSkinMeshObj != static_cast<int>(isSkinMeshObj))
 		{
 			m_cb0_Obj.Work().IsSkinMeshObj = isSkinMeshObj;
 			m_dirtyCBObj = true;
@@ -247,7 +247,7 @@ private:
 	KdConstantBuffer<cbObject>		m_cb0_Obj;				// オブジェクト単位で更新
 	KdConstantBuffer<cbMesh>		m_cb1_Mesh;				// メッシュ毎に更新
 	KdConstantBuffer<cbMaterial>	m_cb2_Material;			// マテリアル毎に更新
-	KdConstantBuffer<cbBone>		m_cb3_Bone;				// スキンメッシュ対応
+	KdConstantBuffer<cbBone>		m_cb3_Bone;				// ボーン毎に更新(スキンメッシュ対応)
 
 	KdRenderTargetPack	m_depthMapFromLightRTPack;
 	KdRenderTargetChanger m_depthMapFromLightRTChanger;
